quotient: NumEvents and NumBranchialEdges from class multiplicities
Under quotient exploration the raw event and branchial counts came from the per-instance
replay, which builds one instance per raw state. On {{1,1},{1,1}} -> {{1,1},{1,1},{1,1}} from
{{1,1},{1,1}} there is one class per step and about 60x more raw states per step: HGEvolve
"NumEvents" took 25.5 s at depth 7 (206,931,038 events) and did not finish depth 8 in 120 s,
while "NumStates" took 0.03 s.

A class c at depth d stands for m(c, d) raw states, and every raw state of c has the matches of
c's expanded representative in the class's slot frame. So m(root, 0) = 1,
m(c', d+1) = sum over matches of c into c' of m(c, d), the raw events are
sum over d < steps of m(c, d) M(c), and the branchial pairs are sum of m(c, d) B(c), with M(c)
the captured matches and B(c) the pairs of them whose consumed slots overlap.
hgcommon/quotient_multiplicity_core.hpp computes this online, one body for both engines:
qm_pass moves a match's unpassed mass on by compare-and-swap on consumed_j(depth), so each unit
of mass crosses each match once whichever side arrives first; qm_credit adds mass and queues the
(class, depth) point unless it is queued; qm_drain runs queued points shallowest first, so one
cascade runs each point once. b_j, counted at capture over the matches linked before j, makes
each overlapping pair count once. The new rendezvous is rv::QuotientMassMatch. On the device
the two entry points are __noinline__: inlined into the persistent kernel, ptxas ran out of its
4 GB cap on persistent.cu. Counts saturate
at 2^63 - 1 with the warning kind "CountSaturated", which the kernel reports as an advisory.

RecordSet::raw_counts_only selects it; the engine uses it when the run reconstructs and the
causal relation is off (Hypergraph::QmCtx on the host, DeviceQmCtx on the device, where the
cascade queue is the driver's descent slice). Both job paths set it through
hgmarshal::reads_raw_counts_only: no Events, EventsMinimal, BranchialEdges, branchial state
edges or event graph requested. Sessions record everything and keep the replay.
qr_run_signature and qr_consumed_overlap are extracted from qr_apply so the replay and the
multiplicity path share them; the device shares qe_frame_step and qe_note_signature between its
two contexts. EvolveResult's reconstructed_raw_events, reconstructed_events and
reconstructed_branchial are 64-bit.

Gates:
- OracleCorpus.MultiplicityCountsMatchTheReplay: states, events and branchial pairs equal the
  replay's on every corpus case under quotient/None at 1 and 8 threads, quotient/Automatic,
  the Automatic reconstruction without quotient exploration, continued in two calls, and
  sampled at rate 0.25; 40 repeats, and 30 under HG_RENDEZVOUS_CHAOS. Counting every earlier
  match as overlapping fails it in 59 cells.
- OracleCorpus.MultiplicityCountsWithoutRawStates: the self-loop rule's closed form, 11 states,
  146,181,741,036,638 events and 2,701,668,796,795,399 branchial pairs at depth 10;
  3,002,019,319,241,196,638 events and a saturated branchial count at depth 12.
- GPU: RecordSet.MultiplicityCountsMatchTheReplayAndTheHost and
  RecordSet.MultiplicityCountsWithoutRawStates (same closed form); gpu_differential_tests
  47/47, hg_gpu_tests 93/93.
- all_tests 326/326; verify_sessions.wls 41/41; verify_paclet.wls failed NONE.
- HGEvolve with the Windows MSVC worker from this tree: depths 5 to 11 match the closed form,
  depth 12 in 0.01 s, depth 13 and 20 saturate with the CountSaturated warning; depth 6 with
  "NumCausalEdges" still takes the replay (3,718,238 events, 4,138,140 causal edges).
- The changed host files pass x86_64-w64-mingw32-g++-posix -fsyntax-only.

Measurement-inert: raw_counts_only defaults off, and no benchmark instrument sets it. With it
off, the only change on a measured path is the extraction of the overlap test and the signature
from qr_apply: callgrind Ir, bench_cpu_evolve 1 thread, HEAD against this tree, wpp depth 6
261,647,128 -> 261,952,195 (+0.12%), multirule depth 6 279,814,786 -> 280,647,045 (+0.30%),
disc2x2 depth 6 1,318,490,212 -> 1,313,289,043 (-0.39%).

+++ gpu/include/hg_gpu/evolve.hpp
@@ -196,7 +196,7 @@ struct EvolveResult {
 
     // Raw events the per-instance replay minted: one per (instance, match) application. The
     // host's num_raw_events under the reconstruction. 0 when the route is off.
-    uint32_t reconstructed_raw_events = 0;
+    uint64_t reconstructed_raw_events = 0;
 
     // Frame alignment: slots the class frame MOVED off the recording state's own labelling, and
     // slots for which no frame image existed (the capture is dropped, so events reachable only
@@ -208,7 +208,7 @@ struct EvolveResult {
 
     // The reconstruction's event count: distinct identities under the run's mode, or the raw
     // application count under EVENT_SIG_NONE, where every application is its own event.
-    uint32_t reconstructed_events = 0;
+    uint64_t reconstructed_events = 0;
 
     // The reconstruction's causal relation: distinct (producer, consumer) pairs, and the
     // consumed-edge occurrences behind them. 0 when the route is off.
@@ -218,7 +218,7 @@ struct EvolveResult {
     // overlap. 0 when the route is off.
     // Pairs tagged in-reduction: the TR view of the same relation.
     uint32_t reconstructed_causal_pairs_reduced = 0;
-    uint32_t reconstructed_branchial = 0;
+    uint64_t reconstructed_branchial = 0;
 
     // The reconstructed relations as pairs of content triples hash(input class, output class,
     // rule) -- the identity a cross-engine comparison is made on, since raw event ids are
